fix(windows): restaure le cast u64 sur RawWindowHandle::Xlib(h.window)
Regression de mon nettoyage clippy precedent : clippy --fix avait
retire le cast `as u64` en signalant `unnecessary_cast` sur Linux,
mais sur Windows `h.window` est u32 (vs u64 sur Linux 64 bits) - le
cast est obligatoire pour le build cross-platform.

CI Windows cassait sur :
  error[E0308]: mismatched types
    expected `u64`, found `u32`

Le cast est restaure avec un `#[allow(clippy::unnecessary_cast)]`
local + un commentaire explicite pour eviter qu'un futur clippy --fix
re-retire le cast.

+++ src-tauri/src/native_player/platform.rs
@@ -38,8 +38,11 @@ impl ParentHandle {
         match handle.as_raw() {
             RawWindowHandle::Win32(h) => ParentHandle::Win32(h.hwnd.get() as u64),
             // h.window est `c_ulong` : u64 sur Linux 64 bits, u32 sur Windows. On caste
-            // explicitement pour eviter une erreur de compile cross-platform.
-            RawWindowHandle::Xlib(h) => ParentHandle::Xlib(h.window),
+            // explicitement pour eviter une erreur de compile cross-platform. Ne PAS
+            // retirer ce cast sous pretexte d'un warning clippy `unnecessary_cast` sur
+            // Linux : le build Windows en a besoin.
+            #[allow(clippy::unnecessary_cast)]
+            RawWindowHandle::Xlib(h) => ParentHandle::Xlib(h.window as u64),
             RawWindowHandle::Xcb(h) => ParentHandle::Xcb(h.window.get() as u64),
             _ => ParentHandle::Unsupported,
         }
